rsg-canoe qol improvements
added a water check so the player has to be standing in the water to launch the canoe.  As of right now player can use canoe 20 feet away from river and float  in the air ,   This fix gets rid of that.
Switched notifys over to ox
added new notify for if you are not standing in the water.